Use "$metadata" as event type for setting stream metadata
This is a system event type and should use this definition in order to
successfully write updates to stream metadata.

It appears to have been defined since at least 1.x, so this should be
safe for most users of this package. It's possible only more recent
versions have enforced the event type.

See also:
https://github.com/EventStore/EventStore/blob/5b3d4b0680e938e497673146eaafe290f53726a3/src/EventStore.ClientAPI/Common/SystemNames.cs#L110-L113